requirements.txt
langchain[openai]==0.0.184
python-dotenv
tabulate
